feat: facilities, dashboard endpoints, and balance integration (Phases 2–5)
Phase 2 — Facility system:
- FacilityType enum (training_pitch, medical_centre, medical_network, scouting_network)
- Facility entity with level 0–5, upgrade costs, and currentEffect descriptions
- FacilityRepository + FacilityService (getAcademyFacilitiesData, upgradeFacility, initializeFacilities)
- AcademyInitializationService now seeds all 4 facility types at level 0
- Academy gets facilities OneToMany collection, hasDebt(), getDebtAmount()

Phase 3 — Dashboard API endpoints:
- GET /api/academy/status — balance, reputation, counts, debt flag
- GET /api/squad — full player list with morale, personality, agent
- GET /api/staff — staff list with role, specialty, morale, salary

Phase 4 — Facility API endpoints:
- GET /api/facilities — list all academy facilities with upgrade info
- POST /api/facilities/{type}/upgrade — deduct cost and increment level

Phase 5 — Balance integration:
- SyncService: earningsDelta → balance, weekly salary deductions, sponsor payments
- SyncService: balance and hasDebt in sync response
- InboxService: typed SponsorRepository/InvestorRepository; idempotency guard;
  signing bonus on sponsor accept; capital injection on investor accept
- EconomicService: investor payouts deducted from academy balance at year-end
- Sponsor: lastPaymentAt + isPaymentDue() for monthly payment tracking
- Migrations: Version20260303210001 (facility table, sponsor.last_payment_at),
  Version20260303201455 (Doctrine schema cleanup)

+++ src/Dto/SyncRequest.php
@@ -13,12 +13,12 @@ class SyncRequest
     public string $clientTimestamp;
 
     #[Assert\PositiveOrZero]
-    public int $earningsDelta;
+    public float $earningsDelta;
 
-    public int $reputationDelta = 0;
+    public float $reputationDelta = 0;
 
     #[Assert\PositiveOrZero]
-    public int $hallOfFamePoints = 0;
+    public float $hallOfFamePoints = 0;
 
     public array $transfers = [];
 }
